增加跳转至证书申请页面时的工时验证，达到工时要求才可以申请证书

+++ app/src/main/java/cn/zheteng123/m_volunteer/entity/user_center/VolunteerEntity.java
@@ -21,7 +21,7 @@ public class VolunteerEntity {
     private String email;
     private String name;
     private int age;
-    private int workingHours;
+    private double workingHours;
     private String avatar;
 
     public int getId() {
@@ -56,11 +56,11 @@ public class VolunteerEntity {
         this.age = age;
     }
 
-    public int getWorkingHours() {
+    public double getWorkingHours() {
         return workingHours;
     }
 
-    public void setWorkingHours(int workingHours) {
+    public void setWorkingHours(double workingHours) {
         this.workingHours = workingHours;
     }
